refactor: replace components from app to root
