+++ astronomy/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Roboto } from "next/font/google";
 import "./globals.css";
-import { Header } from "./components/Header/Header";
+import { Header } from "../components/Header/Header";
 
 const roboto = Roboto({
   weight: ["400", "700"],
