+++ astronomy/tsconfig.json
@@ -19,7 +19,6 @@
       }
     ],
     "paths": {
-      "@assets/*": ["./assets/*"],
       "@/*": ["./*"]
     }
   },
